+++ string.h
@@ -1,11 +1,11 @@
 #include <stddef.h>
 
-void *memcpy(void *dst, void *src, int n);
-void *memmove(void *dst, void *src, int n);
-void *memset(void *s, int v, int n);
-void *memchr(void *s, int c, int n);
-void *memrchr(void *s, int c, int n);
-int memcmp(char *s1, char *s2, int n);
+void *memcpy(void *dst, void *src, long n);
+void *memmove(void *dst, void *src, long n);
+void *memset(void *s, int v, long n);
+void *memchr(void *s, int c, long n);
+void *memrchr(void *s, int c, long n);
+int memcmp(char *s1, char *s2, long n);
 
 char *strcpy(char *dst, char *src);
 char *strchr(char *s, int c);
@@ -13,7 +13,7 @@ char *strrchr(char *s, int c);
 int strlen(char *s);
 int strcmp(char *s1, char *s2);
 
-char *strncpy(char *d, char *s, int n);
+char *strncpy(char *d, char *s, long n);
 char *strcat(char *d, char *s);
-int strncmp(char *d, char *s, int n);
+int strncmp(char *d, char *s, long n);
 char *strstr(char *s, char *r);
